Move period.py and update imports
