+++ catalogue_graph/src/adapters/transformers/ebsco/production.py
@@ -10,7 +10,7 @@ from pymarc.field import Field
 from pymarc.record import Record
 
 from adapters.transformers.ebsco.parsers.field008 import Field008
-from adapters.transformers.ebsco.parsers.period import parse_period
+from adapters.transformers.marc.parsers.period import parse_period
 from adapters.transformers.utils.text_utils import normalise_label
 from models.pipeline.concept import Concept, Period
 from models.pipeline.identifier import Unidentifiable
